CXP-1348: refacto MapProducts into a service
